requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-mongo-connection
pymongo>=4.3.3,<5.0
sqlparse>=0.4.2
dnspython>=2.3.0
django-cors-headers>=4.0.0
